Move permission update functions to permissions controller
